+++ statistics/forms/login.py
@@ -1,0 +1,9 @@
+from django import forms
+from django.contrib.auth import get_user_model
+
+
+class LoginForm(forms.Form):
+    username = forms.CharField(widget=forms.TextInput(attrs={'class': 'form-control'}))
+    password = forms.CharField(widget=forms.PasswordInput(attrs={'class': 'form-control'}))
+
+User = get_user_model()
